fix(V11) : les 3 erreurs remontees par la verification large
- test_squad_spatial_move_mask : la taille du masque etait comparee au litteral
  1047, faux depuis les 20 slots de melee de §9 P3-1. Elle est desormais DERIVEE
  de la composition de l'espace micro ; la valeur absolue reste verrouillee dans
  test_action_space_mirror, ou elle a un sens.
- fight_handlers.model_entry_can_fight_target : engagement_zone annote float
  alors que get_engagement_zone rend un int et que les deux appelants en passent un.
- test_squad_fight_v11_state._fight_action : dict infere dict[str, str], donc
  l'ajout de target_slot (int) ne type-checkait pas — annotation explicite.

+++ engine/phase_handlers/fight_handlers.py
@@ -1670,7 +1670,7 @@ def model_entry_can_fight_target(
     game_state: Dict[str, Any],
     attacker_model_entry: Dict[str, Any],
     target_entry: Dict[str, Any],
-    engagement_zone: float,
+    engagement_zone: int,
 ) -> bool:
     """Coeur de 04.02, sur une empreinte de figurine DEJA construite.
